Replies

+++ apps/client/src/components/comment-form.tsx
@@ -1,22 +1,28 @@
 import { fetchClient } from '@/lib/client';
-import { FeedbackActivitiesQueryData } from '@/lib/query';
+import { FeedbackActivitiesQueryData, FeedbackActivitySummary } from '@/lib/query';
 import { useAuthStore } from '@/stores/auth-store';
 import { useBoolean } from '@paulgeorge35/hooks';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { ChangeEvent, useEffect, useRef, useState } from 'react';
 import { v4 as uuidv4 } from 'uuid';
 import { cn } from '../lib/utils';
+import { Avatar } from './avatar';
 import { Button } from './button';
 import { Icons } from './icons';
 import { ImageFile } from './image-file';
+import { Textbox } from './textbox';
 
 type CommentFormProps = {
     className?: string;
     boardSlug: string;
     feedbackSlug: string;
+    isReply?: boolean;
+    threadId?: string;
+    replyTo?: string;
+    resetReply?: () => void;
 }
 
-export const CommentForm = ({ className, boardSlug, feedbackSlug }: CommentFormProps) => {
+export const CommentForm = ({ className, boardSlug, feedbackSlug, isReply = false, threadId, replyTo, resetReply }: CommentFormProps) => {
     const queryClient = useQueryClient();
     const { user, application } = useAuthStore();
     const [content, setContent] = useState('');
@@ -31,6 +37,7 @@ export const CommentForm = ({ className, boardSlug, feedbackSlug }: CommentFormP
             body: JSON.stringify({
                 content,
                 files: attachments,
+                threadId,
             })
         }),
         onMutate: (content) => {
@@ -43,7 +50,41 @@ export const CommentForm = ({ className, boardSlug, feedbackSlug }: CommentFormP
             queryClient.setQueryData<FeedbackActivitiesQueryData>(['feedback', 'activities', boardSlug, feedbackSlug], (old) => {
                 return {
                     ...old,
-                    activities: [...(old?.activities ?? []), {
+                    activities: threadId ? [
+                        ...(old?.activities ?? []).filter(activity => activity.threadId !== threadId),
+                        {
+                            ...(old?.activities ?? []).find(activity => activity.threadId === threadId),
+                            replies: [
+                                ...(old?.activities ?? []).find(activity => activity.threadId === threadId)?.replies ?? [],
+                                {
+                                    id: 'new',
+                                    author: {
+                                        id: user.id,
+                                        name: user.name,
+                                        avatar: user.avatar,
+                                        isAdmin: user.id === application?.ownerId
+                                    },
+                                    edited: false,
+                                    likes: 0,
+                                    likedByMe: false,
+                                    mergedFromId: null,
+                                    files: attachments.map(attachment => attachment.key),
+                                    createdAt: new Date(),
+                                    updatedAt: new Date(),
+                                    public: true,
+                                    pinned: false,
+                                    type: 'FEEDBACK_COMMENT',
+                                    data: {
+                                        content
+                                    },
+                                    feedbackId: feedbackSlug,
+                                    authorId: user.id,
+                                    threadId: threadId ?? null,
+                                    replies: [] as any,
+                                }
+                            ]
+                        } as FeedbackActivitySummary
+                    ] : [...(old?.activities ?? []), {
                         id: 'new',
                         author: {
                             id: user.id,
@@ -54,6 +95,7 @@ export const CommentForm = ({ className, boardSlug, feedbackSlug }: CommentFormP
                         edited: false,
                         likes: 0,
                         likedByMe: false,
+                        mergedFromId: null,
                         files: attachments.map(attachment => attachment.key),
                         createdAt: new Date(),
                         updatedAt: new Date(),
@@ -64,7 +106,10 @@ export const CommentForm = ({ className, boardSlug, feedbackSlug }: CommentFormP
                             content
                         },
                         feedbackId: feedbackSlug,
-                        authorId: user.id
+                        authorId: user.id,
+                        threadId: null,
+                        thread: null,
+                        replies: [] as any,
                     }]
                 }
             })
@@ -77,6 +122,7 @@ export const CommentForm = ({ className, boardSlug, feedbackSlug }: CommentFormP
         onSuccess: () => {
             setContent('');
             setAttachments([]);
+            resetReply?.();
         },
         onSettled: () => {
             queryClient.invalidateQueries({ queryKey: ['feedback', 'activities', boardSlug, feedbackSlug] });
@@ -121,22 +167,48 @@ export const CommentForm = ({ className, boardSlug, feedbackSlug }: CommentFormP
         },
     })
 
+    useEffect(() => {
+        if (isReply && replyTo) {
+            setContent(` ${replyTo}  `);
+        } else {
+            setContent('');
+        }
+    }, [isReply, replyTo])
+
+    if (isReply && !threadId) {
+        return null;
+    }
+
     return (
-        <div ref={ref} className={cn('flex-1 border', className)}>
-            <div className='p-2'>
-                <input
-                    type="text"
-                    placeholder="Add a comment"
-                    className='w-full focus:outline-none md:text-sm'
-                    value={content}
-                    onChange={(e) => setContent(e.target.value)}
-                    onFocus={expanded.setTrue}
-                    onKeyDown={(e) => {
-                        if (e.key === 'Enter' && !isCommentDisabled) {
-                            comment(content);
-                        }
-                    }}
-                />
+        <div ref={ref} className={cn('flex-1 border vertical gap-1.5', className)}>
+            <div className={cn('p-2', {
+                "pb-[2px]": expanded.value,
+                "p-4": isReply
+            })}>
+                <span className={cn({
+                    "grid grid-cols-[auto_1fr] gap-4": isReply
+                })}>
+                    {isReply && <Avatar
+                        className='size-6 mt-1.5'
+                        src={user?.avatar ?? undefined}
+                        name={user?.name ?? ''}
+                    />}
+                    <Textbox
+                        placeholder="Add a comment"
+                        value={content}
+                        onChange={(e) => setContent(e.target.value)}
+                        wrapperClassName={cn({ 'col-start-2': isReply })}
+                        onFocus={expanded.setTrue}
+                        onKeyDown={(e) => {
+                            if (e.key === 'Enter' && e.metaKey && !isCommentDisabled) {
+                                comment(content);
+                            }
+                        }}
+                        maxLength={1000}
+                        mention={replyTo}
+                        autoFocus={isReply}
+                    />
+                </span>
                 {attachments.length > 0 && <span className='horizontal gap-2 flex-wrap pt-4'>
                     {attachments.map(attachment => (
                         <ImageFile key={attachment.key} fileKey={attachment.key} file={attachment.file} onRemove={() => removeFile(attachment.key)} />
@@ -147,7 +219,7 @@ export const CommentForm = ({ className, boardSlug, feedbackSlug }: CommentFormP
                 'flex': expanded.value
             })}>
                 <input ref={fileInputRef} accept="image/*" name="files" type="file" className="sr-only" multiple onChange={handleFileChange} />
-                <Button size='icon' variant='outline' onClick={() => fileInputRef.current?.click()}>
+                <Button size='icon' variant='outline' disabled={attachments.length >= 3} onClick={() => fileInputRef.current?.click()}>
                     <Icons.Paperclip className="size-4 !stroke-gray-500" />
                 </Button>
                 <button
